pyproject.toml
[project]
name = "codegen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.119.0",
    "openai>=2.3.0",
    "pygithub>=2.8.1",
    "requests>=2.32.5",
    "uvicorn>=0.37.0",
]
